requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
email-validator
python-dotenv
requests
Werkzeug
